Jahresauswahl implementiert
- Datumsauswahl wurde um Jahresauswahl erweitert nicht nur Monatsauswahl.
- Kategorietyp wurde analog zu Dashboard Element Type hochgezogen und abgeändert als Enumeration auch auf Supabase Seite.
- Beim Erstellen und Bearbeiten einer Kategorie wird das Kategoriename Eingabefeld nun automatisch ausgewählt.
- Dashboard Element Karten sind nicht mehr auswählbar auf Dashboard nur noch beim Onboarding.
