+++ lib/data/helper_models/onboarding_models.dart
@@ -7,42 +7,42 @@ import '../models/dashboard_element.dart';
 
 // Start Ausgabekategorien
 final List<StartCategory> expensesStartCategories = [
-  StartCategory(categoryName: 'Lebensmittel', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Haushaltswaren', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Wohnen / Miete', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Nebenkosten', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Bildung', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Kleidung / Mode', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Sport / Fitness', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Freizeit', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Geschenke', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Spende', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Restaurant', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Mobilität', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Abo / Vertrag', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Technik', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Urlaub / Reisen', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Schönheitspflege', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Kredit', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Sonstiges', categoryType: CategoryType.expenses, isSelected: true),
-  StartCategory(categoryName: 'Bar / Kneipe', categoryType: CategoryType.expenses, isSelected: false),
-  StartCategory(categoryName: 'Haustiere', categoryType: CategoryType.expenses, isSelected: false),
-  StartCategory(categoryName: 'Möbel', categoryType: CategoryType.expenses, isSelected: false),
+  StartCategory(categoryName: 'Lebensmittel', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Haushaltswaren', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Wohnen / Miete', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Nebenkosten', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Bildung', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Kleidung / Mode', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Sport / Fitness', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Freizeit', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Geschenke', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Spende', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Restaurant', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Mobilität', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Abo / Vertrag', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Technik', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Urlaub / Reisen', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Schönheitspflege', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Kredit', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Sonstiges', categoryType: CategoryType.expense, isSelected: true),
+  StartCategory(categoryName: 'Bar / Kneipe', categoryType: CategoryType.expense, isSelected: false),
+  StartCategory(categoryName: 'Haustiere', categoryType: CategoryType.expense, isSelected: false),
+  StartCategory(categoryName: 'Möbel', categoryType: CategoryType.expense, isSelected: false),
 ];
 
 // Start Einnahmekategorien
 final List<StartCategory> revenueStartCategories = [
-  StartCategory(categoryName: 'Gehalt', categoryType: CategoryType.revenue, isSelected: true),
-  StartCategory(categoryName: 'Zinsen', categoryType: CategoryType.revenue, isSelected: true),
-  StartCategory(categoryName: 'Weihnachtsgeld', categoryType: CategoryType.revenue, isSelected: true),
-  StartCategory(categoryName: 'Urlaubsgeld', categoryType: CategoryType.revenue, isSelected: true),
-  StartCategory(categoryName: 'Sonstiges', categoryType: CategoryType.revenue, isSelected: true),
-  StartCategory(categoryName: 'Bonus', categoryType: CategoryType.revenue, isSelected: false),
-  StartCategory(categoryName: 'Dividenden', categoryType: CategoryType.revenue, isSelected: false),
-  StartCategory(categoryName: 'Mieteinnahmen', categoryType: CategoryType.revenue, isSelected: false),
-  StartCategory(categoryName: 'Nebeneinkünfte', categoryType: CategoryType.revenue, isSelected: false),
-  StartCategory(categoryName: 'Gewinnprämie', categoryType: CategoryType.revenue, isSelected: false),
-  StartCategory(categoryName: 'Steuerrückerstattung', categoryType: CategoryType.revenue, isSelected: false),
+  StartCategory(categoryName: 'Gehalt', categoryType: CategoryType.income, isSelected: true),
+  StartCategory(categoryName: 'Zinsen', categoryType: CategoryType.income, isSelected: true),
+  StartCategory(categoryName: 'Weihnachtsgeld', categoryType: CategoryType.income, isSelected: true),
+  StartCategory(categoryName: 'Urlaubsgeld', categoryType: CategoryType.income, isSelected: true),
+  StartCategory(categoryName: 'Sonstiges', categoryType: CategoryType.income, isSelected: true),
+  StartCategory(categoryName: 'Bonus', categoryType: CategoryType.income, isSelected: false),
+  StartCategory(categoryName: 'Dividenden', categoryType: CategoryType.income, isSelected: false),
+  StartCategory(categoryName: 'Mieteinnahmen', categoryType: CategoryType.income, isSelected: false),
+  StartCategory(categoryName: 'Nebeneinkünfte', categoryType: CategoryType.income, isSelected: false),
+  StartCategory(categoryName: 'Gewinnprämie', categoryType: CategoryType.income, isSelected: false),
+  StartCategory(categoryName: 'Steuerrückerstattung', categoryType: CategoryType.income, isSelected: false),
 ];
 
 // Startkonten
