+++ lib/features/categories/presentation/pages/update_category_page.dart
@@ -136,12 +136,16 @@ class _UpdateCategoryPageState extends State<UpdateCategoryPage> {
                               _selectedCategoryType = updatedCategoryType;
                             });
                           },
-                          leftValue: CategoryType.expenses,
-                          rightValue: CategoryType.revenue,
-                          leftText: CategoryType.expenses.name,
-                          rightText: CategoryType.revenue.name,
+                          leftValue: CategoryType.expense,
+                          rightValue: CategoryType.income,
+                          leftText: CategoryType.expense.pluralName,
+                          rightText: CategoryType.income.pluralName,
                         ),
-                        TitleInputField(titleController: _categoryNameController, text: 'category_name'),
+                        TitleInputField(
+                          titleController: _categoryNameController,
+                          text: 'category_name',
+                          autoFocus: true,
+                        ),
                         SizedBox(height: 30.0),
                         Hero(
                           tag: 'update_category_fab',
